Move bash configuration to bashrc
The bashrc file is then sourced from bash_profile to provide more
consistent behaviour. The specific issue this solved was that nvim's
`:term` command did not run a login shell by default, therefore I had to
run `:term bash --login` to get my custom shell. This normalises this in
this and other situations.