Fix and improve neovim rake tasks
Due to a breaking change in NeoVim, the configuration files now need to
reside in ./config/nvim and the nvimrc in ./config/nvim/init.vim.

Personally, I want to keep the files where they are (which I find to be
a much more natural place) so I have now modified the install task to
take care of symlinking the already in place files into the new
locations. I made rake nvim:install dependent on the link task to make
sure that the previous symlinks are already in place.

A couple of extra fixes snuck in too:

- Use File.exist? instead, File.directory? didn't work for some reason
  but it should never matter.

- Upgrade the Python module as well in the nvim:update task.