Refactor the Dotfile class to handle all symlinks
It got a second optional argument with the destination path, allowing it
to handle the symlinks that did not go into the home directory (and
previously had their own rake tasks).

I also switched from a blacklist to a whitelist approach to listing the
files, the blacklist was getting fairly long so it seemed reasonable to
simply list what I did want to happen instead of what I didn't.

When a symlink exists, it now also displays the path that is taken so it
can be easily copy pasted to check/delete (since it is not always ~/ now)

This simplified both the Rakefile (by removing some tasks) and the
Dotfile class.