Exit tmux copy-mode when clearing screen with VTR
This solves the problem of running tests, switching to the pane with the
results and scrolling up, then going back to Vim and running them again
only to find out that you are still in copy mode and have to go back to
the old window, exit copy mode, switch panel again and then finally run
the command.

This only adds a "q" to the default clear sequence, this will quit
copy-mode but otherwise actually type it and be cleared away by the
following character.