Add more natural mappings for Vim H and L
The ^ and $ motions are VERY common and bound to keys that are very hard
to press, and on top of that on the wrong side of the keyboard depending
on what they do.

They make perfect sense to bind to the shifted versions of left/right
movements, and the normal bindings which are still useful but not nearly
as often are still available under the "go High" and "go Low" mnemonics.