Scroll without changing panes in tmux
This makes it behave the same in a vim split and a non-vim split,
previously, since Vim has its own scrolling implementation, if you were
hovering over vim it would scroll without changing a pane, but in bash
it would change to that pane.

Only this line:

set -g @scroll-without-changing-pane "on"

...is actually new, the other ones just state the default settings that
were already active, but I find it nicer to have them here for clarity.