Don't fuck up commit diff highlighting after save
I don't quite understand why this is happening, but running the regexes
on the COMMIT_EDITMSG file makes the highlighting of added/removed lines
not work anymore. Perhaps the highlighting is added in a different way
which is no longer present after the highlighting is recalculated after
the commands or something...? Anyhow, I'm just disabling it for now.