git: Remove email from gitconfig
This may be a very simple solution to not accidentally use the work
email privately and vice versa.

It forces you to set it on every repo, which might get annoying. If
that's the case I'll come back to do some includeIf magic - but my bet
is that cloning new repos is infrequent enough, and setting the email
manually is easy enough that it will just always work.

No clever configuration, I'll always know which email to use no matter
what the directory happens to be called on the computer I'm using at
that moment.